// src/generated/models/ServiceImage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ServiceImage = {
    id: string;
    serviceId: string;
    key: string;
    url: string;
    alt?: string | null;
    isPrimary: boolean;
    position: number;
    createdAt: string;
};